misc(bpf-common): split buffer code in separate header
