+++ bpf-common/include/common.bpf.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "vmlinux.h"
 #include <bpf/bpf_core_read.h>
 #include <bpf/bpf_endian.h>
