+++ modules/file-system-monitor/probe.bpf.c
@@ -1,9 +1,8 @@
 // SPDX-License-Identifier: GPL-2.0
+#include "buffer.bpf.h"
 #include "common.bpf.h"
 
 char LICENSE[] SEC("license") = "GPL";
-
-int my_pid = 0;
 
 #define FILE_CREATED 0
 #define FILE_DELETED 1
@@ -13,21 +12,6 @@ int my_pid = 0;
 #define FILE_LINK 5
 #define FILE_RENAME 6
 #define MAX_PATH_COMPONENTS 10
-
-#define BUFFER_MAX 16384
-struct buffer {
-  u32 len;
-  // NOTE: this must be a an array of u64, else we'll have alignment
-  // problems with the Rust code, where the array is a separate struct.
-  // Making this u64 instead of char will result in some alignment padding
-  // added, which is compatible with the Rust one.
-  u64 buffer[BUFFER_MAX / sizeof(u64)];
-};
-
-struct buffer_index {
-  u16 start;
-  u16 len;
-};
 
 struct file_opened_event {
   struct buffer_index filename;
@@ -93,37 +77,6 @@ static struct event_t *init_event(int event_type) {
   return event;
 }
 
-// Copy up to len bytes from source to the buffer pointed by index. Update
-// index.
-static __always_inline void buffer_append_str(struct buffer *buffer,
-                                              struct buffer_index *index,
-                                              const char *source, int len) {
-  // TODO: remove +1 for 0
-  int pos = (index->start + index->len) & (BUFFER_MAX / 2 - 1);
-  if (len > (BUFFER_MAX / 2 - 1))
-    len = BUFFER_MAX / 2 - 1;
-  else
-    len = (len + 1) & (BUFFER_MAX / 2 - 1) /*+6*/;
-  LOG_DEBUG("len = %d", len);
-  // if (len > (BUFFER_MAX / 2 - 1))
-  //   len = (BUFFER_MAX / 2 - 1);
-  char *buf = (char *)buffer->buffer;
-  int r = bpf_core_read_str(&buf[pos], len, source);
-  if (r <= 0) {
-    LOG_ERROR("redding failure: %d", r);
-    return;
-  }
-  LOG_DEBUG("component: %s (+%d)", buf, r);
-  index->len += r - 1;
-  buffer->len += r - 1;
-}
-
-static void buffer_index_init(struct buffer *buffer,
-                              struct buffer_index *index) {
-  index->start = buffer->len;
-  index->len = 0;
-}
-
 // get_path_str was copied and adapted from Tracee
 // Returns the length of the copied entry
 static void get_path_str(struct dentry *dentry, struct path *path,
@@ -135,12 +88,14 @@ static void get_path_str(struct dentry *dentry, struct path *path,
   struct mount *mnt_parent_p = BPF_CORE_READ(mnt_p, mnt_parent);
 
   struct qstr components[MAX_PATH_COMPONENTS];
+#pragma unroll
   for (int i = 0; i < MAX_PATH_COMPONENTS; i++) {
     components[i].len = 0;
     components[i].name = 0;
   }
 
   int count = 0;
+#pragma unroll
   for (int i = 0; i < MAX_PATH_COMPONENTS; i++) {
     struct dentry *mnt_root = NULL;
     mnt_root = (struct dentry *)BPF_CORE_READ(vfsmnt, mnt_root);
@@ -180,14 +135,11 @@ static void get_path_str(struct dentry *dentry, struct path *path,
     buffer_append_str(buffer, index, (char *)components[i].name,
                       components[i].len);
   }
-
-  // LOG_DEBUG("total_len: %d", total_len);
   return;
 }
 
 static __always_inline void output_event(void *ctx, struct event_t *event) {
   unsigned int len = sizeof(struct event_t) - BUFFER_MAX + event->buffer.len;
-  LOG_DEBUG("copied event len: %d/%d", len, event->buffer.len);
   if (len > 0 && len <= sizeof(struct event_t)) {
     bpf_perf_event_output(ctx, &events, BPF_F_CURRENT_CPU, event,
                           len & (BUFFER_MAX - 1));
@@ -203,7 +155,6 @@ static __always_inline void on_path_mknod(void *ctx, struct path *dir,
   if (!event)
     return;
   get_path_str(dentry, dir, &event->buffer, &event->created);
-  // LOG_DEBUG("create %s", event->buffer);
   output_event(ctx, event);
 }
 
@@ -214,7 +165,6 @@ static __always_inline void on_path_unlink(void *ctx, struct path *dir,
   if (!event)
     return;
   get_path_str(dentry, dir, &event->buffer, &event->deleted);
-  // LOG_DEBUG("unlink: %s", event);
   output_event(ctx, event);
 }
 
@@ -226,7 +176,6 @@ static __always_inline void on_file_open(void *ctx, struct file *file) {
   struct path path = BPF_CORE_READ(file, f_path);
   get_path_str(path.dentry, &path, &event->buffer, &event->opened.filename);
   event->opened.flags = BPF_CORE_READ(file, f_flags);
-  // LOG_DEBUG("open %s", event->buffer);
   output_event(ctx, event);
 }
 
@@ -241,7 +190,6 @@ static __always_inline void on_path_link(void *ctx, struct dentry *old_dentry,
   get_path_str(new_dentry, new_dir, &event->buffer, &event->link.source);
   get_path_str(old_dentry, new_dir, &event->buffer, &event->link.destination);
   event->link.hard_link = true;
-  // LOG_DEBUG("hardlink %s", event->buffer);
   //  event->link.destination);
   output_event(ctx, event);
 }
@@ -256,12 +204,9 @@ static __always_inline void on_path_symlink(void *ctx, struct path *dir,
     return;
   get_path_str(dentry, dir, &event->buffer, &event->link.source);
   buffer_index_init(&event->buffer, &event->link.destination);
-  LOG_DEBUG("symlink %d", event->link.destination.start);
   buffer_append_str(&event->buffer, &event->link.destination, old_name,
                     BUFFER_MAX);
-  LOG_DEBUG("symlink %d", event->link.destination.len);
   event->link.hard_link = false;
-  // LOG_DEBUG("symlink %s -> %s", event->link.source, event->link.destination);
   output_event(ctx, event);
 }
 
@@ -273,7 +218,6 @@ static __always_inline void on_path_mkdir(void *ctx, struct path *dir,
   if (!event)
     return;
   get_path_str(dentry, dir, &event->buffer, &event->dir_created);
-  // LOG_DEBUG("mkdir %s", event->buffer);
   output_event(ctx, event);
 }
 
@@ -285,7 +229,6 @@ static __always_inline void on_path_rmdir(void *ctx, struct path *dir,
     return;
   get_path_str(dentry, dir, &event->buffer, &event->dir_deleted);
   event->event_type = DIR_DELETED;
-  // LOG_DEBUG("mkdir %s", event->buffer);
   output_event(ctx, event);
 }
 
@@ -301,7 +244,5 @@ static __always_inline void on_path_rename(void *ctx, struct path *old_dir,
     return;
   get_path_str(old_dentry, old_dir, &event->buffer, &event->rename.source);
   get_path_str(new_dentry, new_dir, &event->buffer, &event->rename.destination);
-  // LOG_DEBUG("rename %s -> %s", event->rename.source,
-  // event->rename.destination);
   output_event(ctx, event);
 }
